cpufreq: interactive: Reschedule timer if min_freq is reduced
When a CPU is running at policy->min, slack timer will not be scheduled.
If policy->min is reduced later, current implementation doesn't
reschedule slack timer and thus could leave CPU at a higher
frequency indefinitely as long as the CPU is idle. This behavior is
undesirable from power perspective.

Change-Id: I40bfd7c93ad3fd06e3837dc48befdc07f29c78c8
Signed-off-by: Junjie Wu <junjiew@codeaurora.org>
Signed-off-by: H34THCL1FF <KayabaAkihikoS4O@gmail.com>